new-minimal-call/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
retell-sdk
httpx==0.27.0
pydantic==2.7.1
